Drag on canvas without selecting
Adds support for dragging objects on the canvas even if they are not
selected -> this works faster as the inspector doesn’t need to load the
object.
