+++ Source/Editor/AGS.Editor/GameCanvas.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using AGS.API;
 using AGS.Engine;
+using Autofac;
+using GuiLabs.Undo;
 
 namespace AGS.Editor
 {
@@ -10,6 +12,7 @@ namespace AGS.Editor
         private readonly IObject _selectionMarker;
         private readonly GameToolbar _toolbar;
         private readonly GameDebugTree _tree;
+        private DragHandle _dragHandle;
 
         public GameCanvas(AGSEditor editor, GameToolbar toolbar, GameDebugTree tree)
         {
@@ -27,6 +30,7 @@ namespace AGS.Editor
         public void Init()
         {
             _editor.Editor.Events.OnRepeatedlyExecute.Subscribe(onRepeatedlyExecute);
+            _editor.Editor.Input.MouseDown.Subscribe(onMouseDown);
             _editor.Editor.Input.MouseUp.Subscribe(onMouseUp);
         }
 
@@ -66,12 +70,32 @@ namespace AGS.Editor
             _selectionMarker.Visible = true;
         }
 
+        private void onMouseDown(MouseButtonEventArgs args)
+        {
+            if (args.Button != MouseButton.Left) return;
+            if (!_selectionMarker.Visible) return;
+            var obj = _editor.Game.HitTest.ObjectAtMousePosition;
+            if (obj == null) return;
+
+            _dragHandle?.Dispose();
+            var handle = _editor.Editor.Factory.Object.GetObject($"{obj.ID}_DraggingHandle");
+            _dragHandle = new DragHandle(handle, _editor, _editor.Editor.State, _editor.Editor.Input, 
+                                         _editor.EditorResolver.Container.Resolve<ActionManager>(), false);
+            _dragHandle.SetBox(obj.GetComponent<IBoundingBoxComponent>());
+            _dragHandle.SetTranslate(obj.GetComponent<ITranslateComponent>());
+            _dragHandle.SetImage(obj.GetComponent<IImageComponent>());
+            handle.GetComponent<IDraggableComponent>().SimulateMouseDown(args);
+        }
+
         private void onMouseUp(MouseButtonEventArgs args)
         {
             if (args.Button != MouseButton.Left) return;
             if (!_selectionMarker.Visible) return;
             var obj = _editor.Game.HitTest.ObjectAtMousePosition;
             if (obj == null) return;
+            var dragHandle = _dragHandle;
+            _dragHandle?.Dispose();
+            if (dragHandle != null && dragHandle.WasDragged) return;
             _tree.Select(obj);
         }
     }
